Make app organization with stub data
I am designing this app as a standard R package. The app code for the
user interface and visualization is all in the `inst/app` folder. The
data analysis code and everything else is in the `R` folder (functions
in this folder may be exported and made available to the R user).

Building the app as a standard R package, rather than an ad hoc
collection of R files and data, allows standard tools to be used to
install, validate and test the app. R has a wonderful ecosystem for
development and will neatly handle dependency installation for us. We
can also use vignettes and ultimately we can submit the whole thing to
CRAN. All of this will improve the quality, maintainability and
accessibility of the app.