Droplevels -- for some reason the AWS and local differ
The AWS barplots seem to show all the clade factors that are present in
the config, even if they are not present. I added droplevels to the
plotting function ... but this should not be the root problem. Why do
they differ? It is possible that the difference is in the package R
versions, since I haven't been able to upgrade packages on my local
government computer.